solo falta detalles

+++ server/prisma/migrations/20230706023922_init/migration.sql
@@ -58,7 +58,7 @@ CREATE TABLE `MetodoDePago` (
 CREATE TABLE `Producto` (
     `id` INTEGER NOT NULL AUTO_INCREMENT,
     `Nombre` VARCHAR(191) NOT NULL,
-    `Precio` DOUBLE NOT NULL,
+    `Precio` DECIMAL(10, 2) NOT NULL,
     `Cantidad` INTEGER NOT NULL,
     `Estado` BOOLEAN NOT NULL,
     `CategoriaID` INTEGER NOT NULL,
@@ -88,7 +88,7 @@ CREATE TABLE `imagen` (
 CREATE TABLE `FacturaDetalle` (
     `id` INTEGER NOT NULL AUTO_INCREMENT,
     `Cantidad` INTEGER NOT NULL,
-    `Subtotal` DOUBLE NOT NULL,
+    `Subtotal` DECIMAL(10, 2) NOT NULL,
     `Estado` VARCHAR(191) NOT NULL,
     `FacturaID` INTEGER NOT NULL,
     `ProductoID` INTEGER NOT NULL,
@@ -100,7 +100,7 @@ CREATE TABLE `FacturaDetalle` (
 CREATE TABLE `Factura` (
     `id` INTEGER NOT NULL AUTO_INCREMENT,
     `Fecha` DATETIME(3) NOT NULL,
-    `Total` DOUBLE NOT NULL,
+    `Total` DECIMAL(10, 2) NOT NULL,
     `Estado` VARCHAR(191) NOT NULL,
     `UsuarioID` INTEGER NOT NULL,
     `DireccionID` INTEGER NOT NULL,
